refactor tests for control statements
